Move cherry-picked branch from dyce into place
...and fix typing errors.
