+++ anydyce/anydice/transformer.py
