+++ anydyce/anydice/unparser.py
@@ -92,7 +92,9 @@ _BINOP_PREC: dict[str, int] = {
 
 
 def unparse(program: Program) -> str:
-    r"""Convert a parsed [`Program`][dyce.anydyce.ast_.Program] to canonical source text."""
+    r"""
+    Convert a parsed [`Program`][anydyce.anydice.ast_.Program] to canonical source text.
+    """
     return "\n".join(_stmt(s, 0) for s in program.stmts)
